requirement.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
db-dtypes==0.3.1
google-api-core==2.7.1
google-auth==2.6.0
google-auth-oauthlib==0.5.0
google-cloud-bigquery==2.34.2
google-cloud-bigquery-storage==2.13.0
google-cloud-core==2.2.3
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
joblib==1.1.0
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
pandas-gbq==0.17.3
proto-plus==1.20.3
protobuf==3.19.4
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.8
